refactor(web): split WalletEntryForm into composable + fields subcomponent
UX-5 PR 3/6: WalletEntryForm.vue reduced from 547 → 96 LOC by extracting
the BRAPI-integrated form state, watchers, validation and submit lifecycle
into useWalletEntryForm composable, and moving NFormItem children into
WalletEntryFormFields subcomponent (injects form via WALLET_ENTRY_FORM_KEY).

Refs #646
